# ProFind/Lib/Global/Services/Swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "WebService",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Tags/{id}": {
      "get": {
        "tags": [
          "Tags"
        ],
        "operationId": "GetTag",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Tag"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Tags"
        ],
        "operationId": "PutTag",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Tag"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Tags"
        ],
        "operationId": "DeleteTag",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Tagtemplates": {
      "get": {
        "tags": [
          "Tagtemplates"
        ],
        "operationId": "GetTagtemplates",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Tagtemplate"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Tagtemplates"
        ],
        "operationId": "PostTagtemplate",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Tagtemplate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Tagtemplate"
            }
          }
        }
      }
    },
    "/api/Tagtemplates/{id}": {
      "get": {
        "tags": [
          "Tagtemplates"
        ],
        "operationId": "GetTagtemplate",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Tagtemplate"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Tagtemplates"
        ],
        "operationId": "PutTagtemplate",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Tagtemplate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Tagtemplates"
        ],
        "operationId": "DeleteTagtemplate",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Activity": {
      "type": "object",
      "properties": {
        "idA": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "titleA": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "descriptionA": {
          "maxLength": 500,
          "minLength": 0,
          "type": "string"
        },
        "expectedBeginA": {
          "$ref": "#/definitions/DateOnly"
        },
        "expectedEndA": {
          "$ref": "#/definitions/DateOnly"
        },
        "pictureA": {
          "format": "byte",
          "type": "string"
        },
        "idPj1": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idT1": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idPj1Navigation": {
          "$ref": "#/definitions/Project"
        },
        "idT1Navigation": {
          "$ref": "#/definitions/Tag"
        },
        "activitycomments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Activitycomment"
          }
        },
        "supporttickets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Supportticket"
          }
        }
      }
    },
    "Activitycomment": {
      "type": "object",
      "properties": {
        "idAc": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "commentAc": {
          "maxLength": 500,
          "minLength": 0,
          "type": "string"
        },
        "dateAc": {
          "$ref": "#/definitions/DateOnly"
        },
        "askToCheckChatAc": {
          "type": "boolean"
        },
        "idA1": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idP5": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC5": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idA1Navigation": {
          "$ref": "#/definitions/Activity"
        },
        "idC5Navigation": {
          "$ref": "#/definitions/Client"
        },
        "idP5Navigation": {
          "$ref": "#/definitions/Professional"
        }
      }
    },
    "Admin": {
      "type": "object",
      "properties": {
        "idA": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "nameA": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "emailA": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "telA": {
          "maxLength": 15,
          "minLength": 0,
          "type": "string"
        },
        "passwordA": {
          "maxLength": 64,
          "minLength": 0,
          "type": "string"
        },
        "pictureA": {
          "format": "byte",
          "type": "string"
        },
        "idR1": {
          "format": "int32",
          "type": "integer"
        },
        "idR1Navigation": {
          "$ref": "#/definitions/Rank"
        },
        "securityansweradmins": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Securityansweradmin"
          }
        },
        "supporttickets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Supportticket"
          }
        }
      }
    },
...
    "Changepasswordcode": {
      "type": "object",
      "properties": {
        "idCpc": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "codeCpc": {
          "maxLength": 64,
          "minLength": 0,
          "type": "string"
        },
        "verifiedCpc": {
          "type": "boolean"
        },
        "idC1": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC1Navigation": {
          "$ref": "#/definitions/Client"
        }
      }
    },
    "Client": {
      "type": "object",
      "properties": {
        "idC": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "nameC": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "emailC": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "passwordC": {
          "maxLength": 64,
          "minLength": 0,
          "type": "string"
        },
        "pictureC": {
          "format": "byte",
          "type": "string"
        },
        "activitycomments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Activitycomment"
          }
        },
        "changepasswordcodes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Changepasswordcode"
          }
        },
        "messages": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Message"
          }
        },
        "projectpays": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Projectpay"
          }
        },
        "projects": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Project"
          }
        },
        "proposalnotifications": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Proposalnotification"
          }
        },
        "proposals": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Proposal"
          }
        },
        "securityanswerclients": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Securityanswerclient"
          }
        },
        "supporttickets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Supportticket"
          }
        }
      }
    },
    "DateOnly": {
      "type": "object",
      "properties": {
        "year": {
          "format": "int32",
          "type": "integer"
        },
        "month": {
          "format": "int32",
          "type": "integer"
        },
        "day": {
          "format": "int32",
          "type": "integer"
        },
        "dayOfWeek": {
          "$ref": "#/definitions/DayOfWeek"
        },
        "dayOfYear": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        },
        "dayNumber": {
          "format": "int32",
          "type": "integer",
          "readOnly": true
        }
      }
    },
...
    "Department": {
      "type": "object",
      "properties": {
        "idDp": {
          "format": "int32",
          "type": "integer"
        },
        "nameDp": {
          "maxLength": 30,
          "minLength": 0,
          "type": "string"
        },
        "professionals": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Professional"
          }
        }
      }
    },
    "Message": {
      "type": "object",
      "properties": {
        "idM": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "contentM": {
          "maxLength": 500,
          "minLength": 0,
          "type": "string"
        },
        "imageM": {
          "format": "byte",
          "type": "string"
        },
        "documentM": {
          "format": "byte",
          "type": "string"
        },
        "locationM": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "audioM": {
          "format": "byte",
          "type": "string"
        },
        "dateTimeM": {
          "$ref": "#/definitions/DateOnly"
        },
        "idP4": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC4": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC4Navigation": {
          "$ref": "#/definitions/Client"
        },
        "idP4Navigation": {
          "$ref": "#/definitions/Professional"
        }
      }
    },
    "Notification": {
      "type": "object",
      "properties": {
        "idN": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "titleN": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "descriptionN": {
          "maxLength": 500,
          "minLength": 0,
          "type": "string"
        },
        "dateTimeIssuedN": {
          "$ref": "#/definitions/DateOnly"
        },
        "pictureN": {
          "format": "byte",
          "type": "string"
        },
        "idPj2": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idPj2Navigation": {
          "$ref": "#/definitions/Project"
        }
      }
    },
    "Profession": {
      "type": "object",
      "properties": {
        "idPfs": {
          "format": "int32",
          "type": "integer"
        },
        "namePfs": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "professionals": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Professional"
          }
        }
      }
    },
    "Professional": {
      "type": "object",
      "properties": {
        "idP": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "nameP": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "dateBirthP": {
          "$ref": "#/definitions/DateOnly"
        },
        "emailP": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "passwordP": {
          "maxLength": 64,
          "minLength": 0,
          "type": "string"
        },
        "activeP": {
          "type": "boolean"
        },
        "sexP": {
          "type": "boolean"
        },
        "duip": {
          "maxLength": 15,
          "minLength": 0,
          "type": "string"
        },
        "afpp": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "isssp": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "zipCodeP": {
          "maxLength": 10,
          "minLength": 0,
          "type": "string"
        },
        "salaryP": {
          "format": "float",
          "type": "number"
        },
        "hiringDateP": {
          "$ref": "#/definitions/DateOnly"
        },
        "pictureP": {
          "format": "byte",
          "type": "string"
        },
        "curriculumP": {
          "format": "byte",
          "type": "string"
        },
        "latitudeLocationP": {
          "format": "double",
          "type": "number"
        },
        "longitudeLocationP": {
          "format": "double",
          "type": "number"
        },
        "idPfs1": {
          "format": "int32",
          "type": "integer"
        },
        "idDp1": {
          "format": "int32",
          "type": "integer"
        },
        "idWdt1": {
          "format": "int32",
          "type": "integer"
        },
        "idDp1Navigation": {
          "$ref": "#/definitions/Department"
        },
        "idPfs1Navigation": {
          "$ref": "#/definitions/Profession"
        },
        "activitycomments": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Activitycomment"
          }
        },
        "messages": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Message"
          }
        },
        "projectpays": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Projectpay"
          }
        },
        "projects": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Project"
          }
        },
        "projecttemplates": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Projecttemplate"
          }
        },
        "proposalnotifications": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Proposalnotification"
          }
        },
        "proposals": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Proposal"
          }
        },
        "securityanswerprofessionals": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Securityanswerprofessional"
          }
        },
        "supporttickets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Supportticket"
          }
        }
      }
    },
    "Project": {
      "type": "object",
      "properties": {
        "idPj": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "titlePj": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "descriptionPj": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "picturePj": {
          "format": "byte",
          "type": "string"
        },
        "totalPricePj": {
          "format": "float",
          "type": "number"
        },
        "idPs1": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idP1": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC1": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC1Navigation": {
          "$ref": "#/definitions/Client"
        },
        "idP1Navigation": {
          "$ref": "#/definitions/Professional"
        },
        "idPs1Navigation": {
          "$ref": "#/definitions/Projectstatus"
        },
        "activities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Activity"
          }
        },
        "notifications": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Notification"
          }
        },
        "projectpays": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Projectpay"
          }
        },
        "supporttickets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Supportticket"
          }
        },
        "tags": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Tag"
          }
        }
      }
    },
    "Projectpay": {
      "type": "object",
      "properties": {
        "idPpy": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "percentagePpy": {
          "format": "float",
          "type": "number"
        },
        "conceptPpy": {
          "maxLength": 500,
          "minLength": 0,
          "type": "string"
        },
        "amountPpy": {
          "format": "float",
          "type": "number"
        },
        "currencyPpy": {
          "maxLength": 3,
          "minLength": 0,
          "type": "string"
        },
        "hasLimitDatePpy": {
          "type": "boolean"
        },
        "limitDatePpy": {
          "$ref": "#/definitions/DateOnly"
        },
        "defaultAmountPpy": {
          "format": "float",
          "type": "number"
        },
        "payStatusPpy": {
          "type": "string"
        },
        "idP3": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC3": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idPj3": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC3Navigation": {
          "$ref": "#/definitions/Client"
        },
        "idP3Navigation": {
          "$ref": "#/definitions/Professional"
        },
        "idPj3Navigation": {
          "$ref": "#/definitions/Project"
        },
        "supporttickets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Supportticket"
          }
        }
      }
    },
    "Projectpaytemplate": {
      "type": "object",
      "properties": {
        "idPpt": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "titlePpt": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "descriptionPpt": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "picturePpt": {
          "format": "byte",
          "type": "string"
        },
        "totalPricePpt": {
          "format": "float",
          "type": "number"
        },
        "idPt1": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idPt1Navigation": {
          "$ref": "#/definitions/Projecttemplate"
        }
      }
    },
    "Projectstatus": {
      "type": "object",
      "properties": {
        "idPs": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "namePs": {
          "maxLength": 20,
          "minLength": 0,
          "type": "string"
        },
        "descriptionPs": {
          "maxLength": 150,
          "minLength": 0,
          "type": "string"
        },
        "colorPs": {
          "maxLength": 6,
          "minLength": 0,
          "type": "string"
        },
        "projects": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Project"
          }
        }
      }
    },
    "Projecttemplate": {
      "type": "object",
      "properties": {
        "idPt": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "titlePt": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "descriptionPt": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "picturePt": {
          "format": "byte",
          "type": "string"
        },
        "totalPricePt": {
          "format": "float",
          "type": "number"
        },
        "saveTagsPt": {
          "type": "boolean"
        },
        "saveProjectPaysPt": {
          "type": "boolean"
        },
        "idP1": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idP1Navigation": {
          "$ref": "#/definitions/Professional"
        },
        "projectpaytemplates": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Projectpaytemplate"
          }
        },
        "tagtemplates": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Tagtemplate"
          }
        }
      }
    },
    "Proposal": {
      "type": "object",
      "properties": {
        "idPp": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "titlePp": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "descriptionPp": {
          "maxLength": 500,
          "minLength": 0,
          "type": "string"
        },
        "picturePp": {
          "format": "byte",
          "type": "string"
        },
        "suggestedStart": {
          "$ref": "#/definitions/DateOnly"
        },
        "suggestedEnd": {
          "$ref": "#/definitions/DateOnly"
        },
        "seen": {
          "type": "boolean"
        },
        "revisionStatus": {
          "type": "string"
        },
        "idP3": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC3": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC3Navigation": {
          "$ref": "#/definitions/Client"
        },
        "idP3Navigation": {
          "$ref": "#/definitions/Professional"
        },
        "proposalnotifications": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Proposalnotification"
          }
        },
        "supporttickets": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Supportticket"
          }
        }
      }
    },
    "Proposalnotification": {
      "type": "object",
      "properties": {
        "idPn": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "contentPn": {
          "maxLength": 500,
          "minLength": 0,
          "type": "string"
        },
        "imagePn": {
          "format": "byte",
          "type": "string"
        },
        "actionPn": {
          "type": "string"
        },
        "idPp1": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idP4": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC4": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC4Navigation": {
          "$ref": "#/definitions/Client"
        },
        "idP4Navigation": {
          "$ref": "#/definitions/Professional"
        },
        "idPp1Navigation": {
          "$ref": "#/definitions/Proposal"
        }
      }
    },
...
    "Securityanswerclient": {
      "type": "object",
      "properties": {
        "idSa": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "answerSa": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "idSq1": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC1": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC1Navigation": {
          "$ref": "#/definitions/Client"
        },
        "idSq1Navigation": {
          "$ref": "#/definitions/Securityquestion"
        }
      }
    },
    "Securityanswerprofessional": {
      "type": "object",
      "properties": {
        "idSa": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "answerSa": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "idSq": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idP": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idPNavigation": {
          "$ref": "#/definitions/Professional"
        },
        "idSqNavigation": {
          "$ref": "#/definitions/Securityquestion"
        }
      }
    },
...
    "Supportticket": {
      "type": "object",
      "properties": {
        "idSt": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "titleSt": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "contentSt": {
          "maxLength": 500,
          "minLength": 0,
          "type": "string"
        },
        "imageSt": {
          "format": "byte",
          "type": "string"
        },
        "documentSt": {
          "format": "byte",
          "type": "string"
        },
        "locationSt": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "audioSt": {
          "format": "byte",
          "type": "string"
        },
        "dateTimeSt": {
          "$ref": "#/definitions/DateOnly"
        },
        "ticketStatusSt": {
          "type": "string"
        },
        "suggestedActionSt": {
          "type": "string"
        },
        "responseContentSt": {
          "maxLength": 500,
          "minLength": 0,
          "type": "string"
        },
        "responseImageSt": {
          "format": "byte",
          "type": "string"
        },
        "responseDocumentSt": {
          "format": "byte",
          "type": "string"
        },
        "responseLocationSt": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "responseAudioSt": {
          "format": "byte",
          "type": "string"
        },
        "idP5": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idC5": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idA2": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idPj4": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idAct1": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idPp2": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idPpy1": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idA2Navigation": {
          "$ref": "#/definitions/Admin"
        },
        "idAct1Navigation": {
          "$ref": "#/definitions/Activity"
        },
        "idC5Navigation": {
          "$ref": "#/definitions/Client"
        },
        "idP5Navigation": {
          "$ref": "#/definitions/Professional"
        },
        "idPj4Navigation": {
          "$ref": "#/definitions/Project"
        },
        "idPp2Navigation": {
          "$ref": "#/definitions/Proposal"
        },
        "idPpy1Navigation": {
          "$ref": "#/definitions/Projectpay"
        }
      }
    },
    "Tag": {
      "type": "object",
      "properties": {
        "idT": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "nameT": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "colorT": {
          "format": "byte",
          "maxLength": 3,
          "type": "string"
        },
        "idPj": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idPjNavigation": {
          "$ref": "#/definitions/Project"
        },
        "activities": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Activity"
          }
        }
      }
    },
    "Tagtemplate": {
      "type": "object",
      "properties": {
        "idTt": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "nameTt": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "colorTt": {
          "format": "byte",
          "maxLength": 3,
          "type": "string"
        },
        "idPt1": {
          "maxLength": 21,
          "minLength": 0,
          "type": "string"
        },
        "idPt1Navigation": {
          "$ref": "#/definitions/Projecttemplate"
        }
      }
    }
  }
}